Split email generation to separate function
